Issue #SB-9206 feat: Changes as per Review comment
