+++ service/src/main/resources/application.conf
@@ -556,13 +556,7 @@ SunbirdMWSystem {
             nr-of-instances = 1
             dispatcher = brr-usr-dispatcher
           }
-        "/RequestRouter/*/TextbookTocUploadActor"
-          {
-            router = smallest-mailbox-pool
-            nr-of-instances = 2
-            dispatcher = rr-usr-dispatcher
-          }
-        "/RequestRouter/*/TextBookTocDownloadActor"
+        "/RequestRouter/*/TextbookTocActor"
           {
             router = smallest-mailbox-pool
             nr-of-instances = 2
